+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 
 * Use Github actions instead of Travis
 * Add Python plugin
+* Add wolfram alpha plugin
 
 1.3.3 (2019-08-26)
 ------------------